+++ internal/auth0/organization/resource_connection_test.go
@@ -1,15 +1,12 @@
-package provider
+package organization_test
 
 import (
-	"context"
-	"fmt"
 	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/stretchr/testify/assert"
 
+	"github.com/auth0/terraform-provider-auth0/internal/provider"
 	"github.com/auth0/terraform-provider-auth0/internal/recorder"
 	"github.com/auth0/terraform-provider-auth0/internal/template"
 )
@@ -46,7 +43,7 @@ func TestAccOrganizationConnection(t *testing.T) {
 	httpRecorder := recorder.New(t)
 
 	resource.Test(t, resource.TestCase{
-		ProviderFactories: TestFactories(httpRecorder),
+		ProviderFactories: provider.TestFactories(httpRecorder),
 		Steps: []resource.TestStep{
 			{
 				Config: template.ParseTestName(TestAccOrganizationConnectionCreate, strings.ToLower(t.Name())),
@@ -94,55 +91,4 @@ func TestAccOrganizationConnection(t *testing.T) {
 			},
 		},
 	})
-}
-
-func TestImportOrganizationConnection(t *testing.T) {
-	var testCases = []struct {
-		testName               string
-		givenID                string
-		expectedOrganizationID string
-		expectedConnectionID   string
-		expectedError          error
-	}{
-		{
-			testName:               "it correctly parses the resource ID",
-			givenID:                "org_1234:conn_5678",
-			expectedOrganizationID: "org_1234",
-			expectedConnectionID:   "conn_5678",
-		},
-		{
-			testName:      "it fails when the given ID is empty",
-			givenID:       "",
-			expectedError: fmt.Errorf("ID cannot be empty"),
-		},
-		{
-			testName:      "it fails when the given ID does not have \":\" as a separator",
-			givenID:       "org_1234conn_5678",
-			expectedError: fmt.Errorf("ID must be formated as <organizationID>:<connectionID>"),
-		},
-		{
-			testName:      "it fails when the given ID has too many separators",
-			givenID:       "org_1234:conn_5678:",
-			expectedError: fmt.Errorf("ID must be formated as <organizationID>:<connectionID>"),
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.testName, func(t *testing.T) {
-			data := schema.TestResourceDataRaw(t, newOrganizationConnection().Schema, nil)
-			data.SetId(testCase.givenID)
-
-			actualData, err := importOrganizationConnection(context.Background(), data, nil)
-
-			if testCase.expectedError != nil {
-				assert.EqualError(t, err, testCase.expectedError.Error())
-				assert.Nil(t, actualData)
-				return
-			}
-
-			assert.Equal(t, actualData[0].Get("organization_id").(string), testCase.expectedOrganizationID)
-			assert.Equal(t, actualData[0].Get("connection_id").(string), testCase.expectedConnectionID)
-			assert.NotEqual(t, actualData[0].Id(), testCase.givenID)
-		})
-	}
 }
